Move to inner-class
